Modified Arduino return data indices
Modified the order in which the Arduino returns data. The goal was to
use a common function to plot data read from the Arduino as well as data
read from a previously saved file.

This introduces some complexity and restrictions on the order in which
data should be returned. If required, could move to a system where every
value has a tag identifying what kind of data it is.